% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breakdown}
\alias{breakdown}
\title{Breakdown times of a machine part}
\format{
A vector with days.
}
\usage{
breakdown
}
\description{
Breakdown times of a machine part
}
\examples{
breakdown
}
\keyword{datasets}
