% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calls}
\alias{calls}
\title{Telephone calls to to the Los Angeles 311 Call Center in 2014}
\format{
A data frame with 846,120 rows and 6 columns:
\itemize{
\item \strong{date} Date.
\item \strong{time} Time.
\item \strong{department} Redirected to which department?
\item \strong{zip} Zip code.
\item \strong{service} Service considered.
\item \strong{solved_how} How solved.
}
}
\usage{
calls
}
\description{
Telephone calls to to the Los Angeles 311 Call Center in 2014
}
\examples{
calls
}
\keyword{datasets}
