% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demand_goods}
\alias{demand_goods}
\title{Demand for two goods/products.}
\format{
A data frame with 362 rows and 3 columns:
date, demand and product id.
}
\usage{
demand_goods
}
\description{
Demand for two goods/products.
}
\examples{
demand_goods
}
\keyword{datasets}
