% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lottery}
\alias{lottery}
\title{Houses visited before selling 20 lottery tickets (assuming only one is sold at a time).}
\format{
A vector with houses.
}
\usage{
lottery
}
\description{
Houses visited before selling 20 lottery tickets (assuming only one is sold at a time).
}
\examples{
lottery
}
\keyword{datasets}
