% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{profit}
\alias{profit}
\title{Profits and costs for financial company}
\format{
A data frame with 24546 rows and 9 variables:
Quarter, Channel, Customer ID, Country, Product Line, Revenue, Product Cost,
Customer Service, Cost and Profit.
}
\usage{
profit
}
\description{
Profits and costs for financial company
}
\examples{
profit
}
\keyword{datasets}
