% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world_pop}
\alias{world_pop}
\title{World population}
\format{
A data frame with 376 rows and 2 variables:
\itemize{
\item \strong{Zip} Zip code.
\item \strong{Area} Name of area.
}
}
\source{
Gapminder project (http://gapm.io/dl_pop).
}
\usage{
world_pop
}
\description{
A dataset containing with populations for each country 1800-2100.
}
\examples{
world_pop
}
\keyword{datasets}
