% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zips}
\alias{zips}
\title{Zip numbers for Jutland in Denmark}
\format{
A data frame with 376 rows and 2 variables:
\itemize{
\item \strong{Zip} Zip code.
\item \strong{Area} Name of area.
}
}
\usage{
zips
}
\description{
A dataset containing the zip numbers for Jutland in Denmark.
}
\examples{
zips
}
\keyword{datasets}
